+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/estimate_ind_cqc_filled_posts_by_job_role.py
@@ -129,14 +129,16 @@ def main(
             pl.scan_parquet(estimates_source, low_memory=True)
             .select(list(combined_schema))
             # Add row id index for single-key joining.
-            .with_row_index(name=IndCQC.row_id)
+            .with_row_index(name=IndCQC.id_per_locationid_import_date)
             .with_columns(utils.cast_to_schema(combined_schema))
         )
         estimated_posts_base_lf = full_estimates_lf.select(
-            IndCQC.row_id, *list(transformation_columns)
+            IndCQC.id_per_locationid_import_date, *list(transformation_columns)
         )
         # This will be joined on at the end.
-        metadata_lf = full_estimates_lf.select(IndCQC.row_id, *list(metadata_columns))
+        metadata_lf = full_estimates_lf.select(
+            IndCQC.id_per_locationid_import_date, *list(metadata_columns)
+        )
 
         col_name_map = {
             IndCQC.ascwds_worker_import_date: IndCQC.ascwds_workplace_import_date
@@ -167,7 +169,7 @@ def main(
         # TODO - Filter ASC-WDS worker data.
 
         estimated_job_role_posts_lf = estimated_job_role_posts_lf.with_row_index(
-            IndCQC.expanded_id
+            IndCQC.id_per_locationid_import_date_job_role
         )
 
         # Remove log_polars_plan.
@@ -285,7 +287,7 @@ def main(
         # Join back original metadata.
         estimated_job_role_posts_lf = estimated_job_role_posts_lf.join(
             metadata_lf,
-            on=IndCQC.row_id,
+            on=IndCQC.id_per_locationid_import_date,
         )
 
         # Remove log_polars_plan.
@@ -316,7 +318,9 @@ def join_estimates_to_ascwds(
 
     # Narrow select "id" and join keys first to improve memory performance of
     # cross join. From this we get the full amount of rows expected.
-    narrow_keys_lf = estimates_lf.select([IndCQC.row_id] + join_keys)
+    narrow_keys_lf = estimates_lf.select(
+        [IndCQC.id_per_locationid_import_date] + join_keys
+    )
     # This is just a single column df with a row for each job role (~38).
     roles_lf = pl.LazyFrame(
         data=[AscwdsWorkerValueLabelsJobGroup.all_roles()],
@@ -336,7 +340,7 @@ def join_estimates_to_ascwds(
     # both sides as they are not relevant to the rest of the pipeline.
     return estimates_lf.join(
         expanded_counts_lf.drop(join_keys),
-        on=IndCQC.row_id,
+        on=IndCQC.id_per_locationid_import_date,
         how="right",
     ).drop(join_keys)
 
@@ -363,15 +367,18 @@ def impute_ratios(estimated_job_role_posts_lf: pl.LazyFrame) -> pl.LazyFrame:
         estimated_job_role_posts_lf.group_by(impute_groups)
         .agg(
             # Sort the join key in the same manner as the imputed values.
-            pl.col(IndCQC.expanded_id).sort_by(order_key),
+            pl.col(IndCQC.id_per_locationid_import_date_job_role).sort_by(order_key),
             imputed_ratios,
         )
-        .explode(IndCQC.expanded_id, IndCQC.imputed_ascwds_job_role_ratios)
+        .explode(
+            IndCQC.id_per_locationid_import_date_job_role,
+            IndCQC.imputed_ascwds_job_role_ratios,
+        )
         .drop(impute_groups)
     )
 
     return estimated_job_role_posts_lf.join(
-        impute_agg_lf, on=IndCQC.expanded_id, how="left"
+        impute_agg_lf, on=IndCQC.id_per_locationid_import_date_job_role, how="left"
     )
 
 
@@ -386,20 +393,22 @@ def get_percent_share_ratios(
     """
     groups = [IndCQC.location_id, IndCQC.cqc_location_import_date]
 
-    # Groupby-agg-explode on only necessary subset, before joining back on expanded_id.
+    # Groupby-agg-explode on only necessary subset, before joining back on id_per_locationid_import_date_job_role.
     ratios_agg_lf = (
         estimated_job_role_posts_lf.group_by(groups)
         .agg(
-            pl.col(IndCQC.expanded_id),  # Keep to align during explode
+            pl.col(
+                IndCQC.id_per_locationid_import_date_job_role
+            ),  # Keep to align during explode
             percentage_share(input_col).cast(pl.Float32).alias(output_col),
         )
-        .explode(IndCQC.expanded_id, output_col)
+        .explode(IndCQC.id_per_locationid_import_date_job_role, output_col)
         # Drop groups to prevent duplicate columns after join.
         .drop(groups)
     )
 
     return estimated_job_role_posts_lf.join(
-        ratios_agg_lf, on=IndCQC.expanded_id, how="left"
+        ratios_agg_lf, on=IndCQC.id_per_locationid_import_date_job_role, how="left"
     )