+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -1394,7 +1394,7 @@ class TestJoinEstimatesToAscwds:
     TEST_ROLES = ["role_a", "role_b"]
     estimates_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.ascwds_workplace_import_date: pl.String,
             IndCQC.establishment_id: pl.String,
         }
@@ -1409,7 +1409,7 @@ class TestJoinEstimatesToAscwds:
     )
     expected_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(TEST_ROLES),
             "value": pl.Float64,
         }
@@ -1420,7 +1420,7 @@ class TestJoinEstimatesToAscwds:
 class ImputeJobRoleSchemas:
 
     create_imputed_ascwds_job_role_counts_expected_schema = {
-        IndCQC.expanded_id: pl.UInt32,
+        IndCQC.id_per_locationid_import_date_job_role: pl.UInt32,
         IndCQC.location_id: pl.String,
         IndCQC.main_job_role_clean_labelled: pl.String,
         IndCQC.cqc_location_import_date: pl.Date,
@@ -1432,7 +1432,7 @@ class ImputeJobRoleSchemas:
     }
 
     create_ascwds_job_role_rolling_ratio_expected_schema = {
-        IndCQC.expanded_id: pl.UInt16,
+        IndCQC.id_per_locationid_import_date_job_role: pl.UInt16,
         IndCQC.location_id: pl.String,
         IndCQC.cqc_location_import_date: pl.Date,
         IndCQC.primary_service_type: pl.String,
@@ -1462,7 +1462,7 @@ class ImputeJobRoleSchemas:
 class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
     calculate_estimated_filled_posts_by_job_role_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.estimate_filled_posts: pl.Float32,
             IndCQC.ascwds_job_role_ratios: pl.Float32,
             IndCQC.imputed_ascwds_job_role_ratios: pl.Float32,
@@ -1482,7 +1482,7 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
 
     adjust_managerial_roles_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
                 AscwdsWorkerValueLabelsJobGroup.all_roles()
             ),
@@ -1492,7 +1492,7 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
     )
     expected_adjust_managerial_roles_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
                 AscwdsWorkerValueLabelsJobGroup.all_roles()
             ),
@@ -1503,7 +1503,7 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
 
     expected_calculate_reg_man_difference_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
                 AscwdsWorkerValueLabelsJobGroup.all_roles()
             ),
@@ -1515,7 +1515,7 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
 
     expected_calculate_non_rm_managerial_distribution_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
                 AscwdsWorkerValueLabelsJobGroup.all_roles()
             ),
@@ -1526,7 +1526,7 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
 
     expected_distribute_rm_difference_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
                 AscwdsWorkerValueLabelsJobGroup.all_roles()
             ),
@@ -1540,7 +1540,7 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
 
     expected_calc_diff_estimate_filled_posts_and_from_all_job_roles_schema = pl.Schema(
         {
-            IndCQC.row_id: pl.Int32,
+            IndCQC.id_per_locationid_import_date: pl.Int32,
             IndCQC.estimate_filled_posts: pl.Float32,
             IndCQC.main_job_role_clean_labelled: pl.Enum(
                 AscwdsWorkerValueLabelsJobGroup.all_roles()